Start common endpoints and swagger work.
