+++ auslib/web/common/swagger/definitions.yml
@@ -95,6 +95,18 @@ definitions:
         maxLength: 15
         example: Firefox
 
+  ChannelModel:
+    title: Channel
+    description: "model to define 'channel' field in request body."
+    type: object
+    properties:
+      channel:
+        description: The update channel of the application request an update.
+        type: string
+        minLength: 1
+        maxLength: 75
+        example: beta
+
   ReleaseName:
     title: Release Name
     description: name of release
@@ -400,3 +412,43 @@ definitions:
                 update_type: "minor"
                 version: null
                 whitelist: null
+
+  EmergencyShutoffObject:
+    title: Emergency shut off information.
+    description: Information about product/channel available to shut off updates.
+    type: object
+    required:
+      - product
+      - channel
+    allOf:
+      - $ref: '#/definitions/ProductModel'
+      - $ref: '#/definitions/ChannelModel'
+      - $ref: '#/definitions/DataVersionModel'
+      - type: object
+        properties:
+          shutoff_id:
+            description: Shut off Id
+            type: integer
+            format: int32
+          updates_disabled:
+            description: Indicates if updates are disabled for product/channel.
+            type: boolean
+          additional_notification_list:
+            description: Email notification list, additional to default mailing list.
+            type: ["string", "null"]
+
+  EmergencyShutoffListGET:
+    title: List of emergency shut offs.
+    description: Product/Channels available to shut off updates.
+    type: object
+    required:
+      - shutoffs
+    allOf:
+      - $ref: '#/definitions/CountModel'
+      - type: object
+        properties:
+          shutoffs:
+            description: List of emergency shutoffs.
+            type: array
+            items:
+              $ref: '#/definitions/EmergencyShutoffObject'
